# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MetalMark API
  description: Self-hosted personal finance (Monarch-style). LAN/VPN only.
  version: 0.1.0
paths:
...
  /transactions:
    post:
      tags:
      - transactions
      summary: Create Transaction
      operationId: create_transaction_transactions_post
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - transactions
      summary: List Transactions
      operationId: list_transactions_transactions_get
      parameters:
      - name: account_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: 'null'
          title: Account Id
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: 'null'
          title: Category Id
      - name: owner_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End
      - name: review_status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Hidden
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions/{txn_id}:
    get:
      tags:
      - transactions
      summary: Get Transaction
      operationId: get_transaction_transactions__txn_id__get
      parameters:
      - name: txn_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Txn Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - transactions
      summary: Update Transaction
      operationId: update_transaction_transactions__txn_id__patch
      parameters:
      - name: txn_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Txn Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - transactions
      summary: Delete Transaction
      operationId: delete_transaction_transactions__txn_id__delete
      parameters:
      - name: txn_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Txn Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions/{txn_id}/splits:
    put:
      tags:
      - transactions
      summary: Replace Splits
      operationId: replace_splits_transactions__txn_id__splits_put
      parameters:
      - name: txn_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Txn Id
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/SplitIn'
              title: Splits
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /transactions/transfers:
    post:
      tags:
      - transactions
      summary: Link Transfer
      operationId: link_transfer_transactions_transfers_post
      parameters:
      - name: metalmark_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Metalmark Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferLink'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SplitIn:
      properties:
        amount:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Amount
        pct:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Pct
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: SplitIn
    SplitOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount
        base_amount:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Base Amount
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        effective_owner_id:
          type: string
          format: uuid
          title: Effective Owner Id
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - amount
      - base_amount
      - category_id
      - owner_id
      - effective_owner_id
      - notes
      title: SplitOut
...
    TransactionCreate:
      properties:
        account_id:
          type: string
          format: uuid
          title: Account Id
        amount:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount
        transacted_at:
          type: string
          format: date-time
          title: Transacted At
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        merchant:
          anyOf:
          - type: string
          - type: 'null'
          title: Merchant
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        is_pending:
          type: boolean
          title: Is Pending
          default: false
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        tag_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Tag Ids
          default: []
      type: object
      required:
      - account_id
      - amount
      - transacted_at
      title: TransactionCreate
    TransactionOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        account_id:
          type: string
          format: uuid
          title: Account Id
        amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount
        currency:
          type: string
          title: Currency
        base_amount:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Base Amount
        fx_rate_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Fx Rate Date
        transacted_at:
          type: string
          format: date-time
          title: Transacted At
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        merchant:
          anyOf:
          - type: string
          - type: 'null'
          title: Merchant
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        effective_owner_id:
          type: string
          format: uuid
          title: Effective Owner Id
        is_pending:
          type: boolean
          title: Is Pending
        review_status:
          type: string
          title: Review Status
        is_hidden:
          type: boolean
          title: Is Hidden
        is_split_parent:
          type: boolean
          title: Is Split Parent
        transfer_group_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Transfer Group Id
        field_sources:
          additionalProperties: true
          type: object
          title: Field Sources
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        source:
          type: string
          title: Source
        tag_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Tag Ids
          default: []
        splits:
          items:
            $ref: '#/components/schemas/SplitOut'
          type: array
          title: Splits
          default: []
      type: object
      required:
      - id
      - account_id
      - amount
      - currency
      - base_amount
      - fx_rate_date
      - transacted_at
      - posted_at
      - description
      - merchant
      - category_id
      - owner_id
      - effective_owner_id
      - is_pending
      - review_status
      - is_hidden
      - is_split_parent
      - transfer_group_id
      - field_sources
      - notes
      - source
      title: TransactionOut
    TransactionPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TransactionOut'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - items
      title: TransactionPage
    TransactionUpdate:
      properties:
        amount:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Amount
        transacted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Transacted At
        posted_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Posted At
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        merchant:
          anyOf:
          - type: string
          - type: 'null'
          title: Merchant
        category_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Category Id
        owner_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Owner Id
        is_pending:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Pending
        is_hidden:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Hidden
        review_status:
          anyOf:
          - type: string
            pattern: ^(needs_review|reviewed|ignored)$
          - type: 'null'
          title: Review Status
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        tag_ids:
          anyOf:
          - items:
              type: string
              format: uuid
            type: array
          - type: 'null'
          title: Tag Ids
      type: object
      title: TransactionUpdate
      description: 'Absent means "no change"; explicit null means "clear" for the nullable fields.


        For the rest null is a client bug, and saying so beats ignoring it — silently

        dropping a null is how a PATCH appears to succeed while changing nothing.'
    TransferLink:
      properties:
        from_txn_id:
          type: string
          format: uuid
          title: From Txn Id
        to_txn_id:
          type: string
          format: uuid
          title: To Txn Id
      type: object
      required:
      - from_txn_id
      - to_txn_id
      title: TransferLink
      description: Manually link two existing transactions as a transfer (ADR-0008/0018).
    TransferOut:
      properties:
        transfer_group_id:
          type: string
          format: uuid
          title: Transfer Group Id
        matched_by:
          type: string
          title: Matched By
        fx_cost_base:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Fx Cost Base
        txn_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Txn Ids
      type: object
      required:
      - transfer_group_id
      - matched_by
      - fx_cost_base
      - txn_ids
      title: TransferOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError